ingestion/requirements.txt
pandas>=2.1,<3.0
pydantic>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
boto3>=1.34,<2.0
